wdl   调用视频识别入库接口

+++ rockburst-system/src/main/java/com/ruoyi/system/service/IBizProjectRecordService.java
@@ -32,7 +32,7 @@ public interface IBizProjectRecordService  extends MPJDeepService<BizProjectReco
     MPage<BizProjectRecordListVo> getlist(BasePermission permission, BizProjectRecordDto dto , Pagination pagination);
 
 
-    int saveRecord(BizProjectRecordAddDto dto);
+    Long saveRecord(BizProjectRecordAddDto dto);
 
     int saveRecordApp(BizProjectRecordAddDto dto);
